cli/pyproject.toml
[project]
name = "keboola-duckdb-cli"
version = "0.1.0"
description = "CLI tool for Keboola DuckDB Storage API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Keboola", email = "support@keboola.com"}
]
keywords = ["keboola", "duckdb", "cli", "storage"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "respx>=0.20.0",
]

[project.scripts]
keboola-duckdb = "keboola_duckdb_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/keboola_duckdb_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
